Cleanups in preparation for Editability & Multi-chats
